+++ pkg/kiosk/aws_login.go
@@ -58,6 +58,9 @@ func GrafanaKioskAWSLogin(ctx context.Context, cfg *Config, dir string, messages
 			panic(err)
 		}
 	}
+	if err := chromedp.Run(taskCtx, triggerAutofit(cfg)); err != nil {
+		panic(err)
+	}
 
 	// blocking wait until context is cancelled or a message triggers a reload
 	for {
@@ -67,6 +70,7 @@ func GrafanaKioskAWSLogin(ctx context.Context, cfg *Config, dir string, messages
 		case messageFromChrome := <-messages:
 			if err := chromedp.Run(taskCtx,
 				chromedp.Navigate(generatedURL),
+				triggerAutofit(cfg),
 			); err != nil {
 				return
 			}