+++ pkg/kiosk/grafana_idtoken_login.go
@@ -73,6 +73,9 @@ func GrafanaKioskIDToken(ctx context.Context, cfg *Config, dir string, messages 
 	if err := chromedp.Run(taskCtx, enableFetch(generatedURL)); err != nil {
 		panic(err)
 	}
+	if err := chromedp.Run(taskCtx, triggerAutofit(cfg)); err != nil {
+		panic(err)
+	}
 
 	// blocking wait until context is cancelled or a message triggers a reload
 	for {
@@ -82,6 +85,7 @@ func GrafanaKioskIDToken(ctx context.Context, cfg *Config, dir string, messages 
 		case messageFromChrome := <-messages:
 			if err := chromedp.Run(taskCtx,
 				chromedp.Navigate(generatedURL),
+				triggerAutofit(cfg),
 			); err != nil {
 				return
 			}